requirements.txt
Flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
Pillow==8.3.1
tqdm==4.61.2
redis==3.5.3
anthropic==0.1.0  # Ensure this matches the correct version for your API
google-cloud==0.34.0  # Or the specific Google Cloud libraries you are using
python-dotenv==0.19.0
